requirements.txt
selenium
mysql-connector-python
python-dotenv
beautifulsoup4
requests
